requirements.txt
Django==3.2
djangorestframework==3.12
djangorestframework-simplejwt==4.7
sqlalchemy==1.4
mysqlclient==2.0

